// Commerce-be/src/modules/category/model/dto.ts
import { z } from "zod";
import { ModelStatus } from "../../../share/model/base-model";

export const CategoryCreateSchema = z.object({
  name: z.string().min(2, "name must be at least 2 characters"),
  image: z.string().optional(),
  description: z.string().optional(),
  parentId: z.string().uuid().nullable().optional(),
});
export type CategoryCreateDTO = z.infer<typeof CategoryCreateSchema>;

export const CategoryUpdateSchema = z.object({
  name: z.string().min(2, "name must be at least 2 characters").optional(),
  image: z.string().optional(),
  description: z.string().max(255, "description must be at most 255 characters").optional(),
  parentId: z.string().uuid().nullable().optional(),
  status: z.nativeEnum(ModelStatus).optional(),
});
export type CategoryUpdateDTO = z.infer<typeof CategoryUpdateSchema>;

export const CategoryCondDTOSchema = z.object({
  name: z.string().min(2, "name must be at least 2 characters").optional(),
  parentId: z.string().uuid().nullable().optional(),
  status: z.nativeEnum(ModelStatus).optional(),
});
export type CategoryCondDTO = z.infer<typeof CategoryCondDTOSchema>;